{topost} scsi: libsas: check if the same sata device when flutter
The ata device do not have a real sas address. If a ata device is
replaced with another one, the sas address is the same. Now libsas treat
this senario as flutter and do not delete the old one and discover the
new one. This will cause the date read from or write to the wrong
device.

And also when hotplugging a sata device, libsas entered to the flutter
case and found the phy attached address is abnormal.

Fix this issue by checking the phy attached address and the ata device's
class and id.

Signed-off-by: Jason Yan <yanaijie@huawei.com>