{fromlist v2} HISI LPC: Stop using MFD APIs
The MFD APIs should only be used by drivers in
drivers/mfd. It is not worth splitting the driver
to have separate parts in drivers/bus and
drivers/mfd, so just drop MFD usage.

As a solution, we will use the platform device APIs
directly to achieve the same as we had when using MFD
APIs.

Signed-off-by: John Garry <john.garry@huawei.com>