{topost} scsi: hisi_sas: rename hisi_sas_phy.{is_flutter -> in_reset}
Currently to flag a PHY in reset we use "is_flutter" flag.
Flutter generally means a glitch occurred on the PHY, while
the PHY remains up. This is not really what we use the flag
for, so rename to a more appropriate name.

In addition, hisi_sas_phy elements are rearranged to improve
packing efficiency.

Fixes: 6fb89f96a175 ("scsi: hisi_sas: mark hardreset as PHY flutter for directly attached device")
Signed-off-by: John Garry <john.garry@huawei.com>
Reviewed-by: Xiang Chen <chenxiang66@hisilicon.com>