{topost} scsi: hisi_sas: make nexus reset for local PHY synchronous
Currently we patiently wait for 2000ms for a local PHY to
be reset for a nexus reset. This 2000ms is just a safe
guess for reset completion time.

To improve reliability and speed, make this synchronous by
using a per-PHY completion.

We also remove the handling in the nexus reset code for
when the attached SAS address has changed during the reset.

This code was superfluous, as if the PHY SAS address changes
during the nexus reset, a new SAS address will be reported
to libsas: in this scenario libsas will deform the previous
port formed, and create a new one, thereby de-registering the
old device attached and registering the new one.

In the case where the PHY has not come back up for the
nexus reset, we re-issue the PHY down event to libsas.

Fixes: 6fb89f96a175 ("scsi: hisi_sas: mark hardreset as PHY flutter for directly attached device")
Signed-off-by: John Garry <john.garry@huawei.com>
Reviewed-by: Xiang Chen <chenxiang66@hisilicon.com>