{topost} scsi: hisi_sas: update caller's return value when get_free_slot() failed
It is an step of executing task to get free slot. If the step fails,
we will cleanup LLDD resources and should return failure to upper layer
or internal caller to abort task execution of this time.

But in the current code, the caller of get_free_slot() doesn't return
failure when get_free_slot() failed. This patch is to fix it.

Fixes: 616898da772a ("{topost} scsi: hisi_sas: optimise the dq lock
usage when delivery IO")

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>