Revert "{topost} net: hns3: Add TX_QUEUE_TC cmd support in TM module"
This reverts commit 9419abcedbacf0fee45c31856ef0ee51b511160a.

commit 7b64db823801 ("{topost} net: hns3: clear TX/RX rings when stopping port")

Above commit reset the queue when down/up the netdev, which
also reset the tx ring's tc info.

We revert the patch here, because this patch does not support
restoring the tx ring's tc info when down/up the netdev.

Also, this patch only works for PF, but not work for VF, bacause
ring_id in hclge_tx_ring_to_tc_cmd is relative id in a function.

We will introduce a new patch to set the tx ring's tc info.

Signed-off-by: Yunsheng Lin <linyunsheng@huawei.com>
Signed-off-by: Peng Li <lipeng321@huawei.com>