{topost} net: hns3: add tcam config support for flow director
A flow director rule includes a key and several actions.
The key is combined with a set of packet tuples, stored
in the TCAM block. For each TCAM entry bit, it uses a pair
of bit 'x' and 'y' to indicate which value to match. There
are several actions can be selected when a packet hit the
total key, typically incluede forwarding to a specified
queue or dropping it.

Signed-off-by: Jian Shen <shenjian15@huawei.com>
Signed-off-by: Peng Li <lipeng321@huawei.com>