scsi: libsas: unregister the device if linkrate changed or ata device disabled
When the linkrate is changed after update the phy state in the flutter
case, all commands failed. This is because LLDD did not re-config the
device with the new linkrate. Since LLDD only config the linkrate in
lldd_dev_found() callback, we need to unregister the device and probe it
again.

On the other hand, if a ata device is disabled in libata, libsas does
nothing now. When a new phy up comes, it is considered as flutter and
will not be deleted and probed again.

Signed-off-by: Jason Yan <yanaijie@huawei.com>