irqchip/gic-v3-its: do not set affinity to offline node
In numa system, a node may not available (or offline).

NUMA node will be built if CPUs and (or) memory are valid on this NUMA node;

But if we boot the system with memory-less node and also with CONFIG_NR_CPUS
less than CPUs in SRAT, for example, 64 CPUs total with 4 NUMA nodes, 16 CPUs
on each NUMA node, if we boot with CONFIG_NR_CPUS=48, then we will not built
numa node for node 3.

So, we don't set affinity to a invalid node.

Signed-off-by: Xie XiuQi <xiexiuqi@huawei.com>