{topost} net: hns3: fix for multiple unmapping DMA problem
We split a big fragment to many small fragments and a small fragment
take up a descriptor. But we only have mapped the big fragment. When
doing unmap, we just need to unmap the big fragment instead of every
small fragment. To solve this problem, this patch makes the value of
desc_cb.length of small fragment becoming zero. If desc_cb.length is
zero, we do not unmap the buffer.

Signed-off-by: Fuyun Liang <liangfuyun1@huawei.com>
Signed-off-by: Peng Li <lipeng321@huawei.com>