{frommaintainer} scsi: hisi_sas: fix the failure of recovering PHY from STP link timeout
There is an issue that link reset can't recover PHY when STP
link timeout. Because current process of enabling PHY for v3 hw
will wait last transmission done. The time of one transmission
depends IO size, disk model and so on. Normally, it should be
shorter than 50ms. But the last transmission could be never done
for some abnormal scenarios, such as STP link timeout.

This patch is to fix the issue. Check PHY status after starting
process of enabling PHY for 50ms. If the PHY is still active, we
close it forcibly by PHY reset. Of course, we need to clear the
PHY reset bit when enable PHY.

Besides, the function disable_phy_v3_hw() should not be suitable to
call in interrupts for hilink bug for this 50ms delay. Then, we do link
reset for hilink bug directly. The change is that we don't clear the
invalid dword count register. This is better. Because we should not
clear such error count while not saved.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>