{topost} net: hns3: fix query fd information missing issue
In oringinal codes, the flow_type of fd rule was broken, the flags
FLOW_EXT and FLOW_MAC_EXT were dropped. When query the fd information,
the extra information of the rules were lost. The mac information of
FLOW_MAC_EXT is also incorrect by using the mask value. This patch fixes
them.

Fixes: 80f70308e381 ("{topost} net: hns3: add support for rule add/delete for flow director")
Signed-off-by: Jian Shen <shenjian15@huawei.com>
Signed-off-by: Peng Li <lipeng321@huawei.com>