{topost} net: hns3: refine the handle for hns3_nic_net_open/stop()
When trigger nic down, there is a time window between bring down the
protocol stack and stop the work task. If the net was up in the time
window, it may bring up the protocol stack again.

This patch fixes it by stop the work task at the beginning of
hns3_nic_net_stop(). To keep symmetrical, start the work task at the
end of hns3_nic_net_open().

Signed-off-by: Jian Shen <shenjian15@huawei.com>
Signed-off-by: Peng Li <lipeng321@huawei.com>