{topost} scsi: hisi_sas: add memory barrier in task start delivery function
In task start delivery function, we need add an memory barrier to
prevent re-order of reading memory by hardware. Because the slot
data is set in task prepare function and it could be running in
another CPU.

This patch adds an memory barrier after s->ready read in task start
delivery function, and uses WRITE_ONCE() in the places where s->ready
is set.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>