{topost} RDMA/hns: Limit extend sq sge num
According to hip08 limit, the buffer size of extend sge needs
to be an integer wqe_sge_buf_page size. For example, the value
of sge_shift field of qp context is greater or equal to eight
when buf page size is 4K size. The value of sge_shift field of
qp context assgined by hr_qp->sge.sge_cnt. This patch fixed it.

Signed-off-by: Lijun Ou <oulijun@huawei.com>