{topost} scsi: hisi_sas: set the value of aging timer to minimum value
The aging timer is to wait 4 CQs in limit time, and then report CQs
together to driver. But for serial IO scenario,it will still wait 8us for
every IO before it is reported to driver. So in the situation, the
performance is very low. So to improve it, change the limit time to the
least value.
For other scenario, it does little affect to the performance.

Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Signed-off-by: John Garry <john.garry@huawei.com>