{topost} net: hns3: add flag to indicate enet need to be reinitialized
In order to prevent enet from being repeatedly uninitialized,
hns3_nic_priv->state sets the HNS3_NIC_STATE_REINITING flag to indicate
that enet needs to be reinitialized.
When the HNS3_NIC_STATE_REINITING flag is set, it means that
the corresponding initialization is not completed. So access
to coresponding variables should not be allowed.

Signed-off-by: Huazhong Tan <tanhuazhong@huawei.com>
Signed-off-by: Peng Li <lipeng321@huawei.com>