KVM: arm64: Force guest EL1 to use user-space's partid configuration
While we trap the guests attempts to read/write the MPAM control registers,
these remain in effect. guest-EL0 uses KVM's user-space's configuration,
as the value is left in the register, and guest-EL1 uses either the host
kernel's configuration, or in the case of VHE, the unknown reset value of
MPAM1_EL1.

We want to force the guest-EL1 to use KVM's user-space's MPAM configuration.

When loading the vcpu on a VHE system copy the EL0 register value from
context-switch into the EL1 register. For non-VHE we do the same at EL2,
and on guest-exit we restore the host configuration to EL1, which we know
is 0.

Signed-off-by: James Morse <james.morse@arm.com>