DTS:DTS2019030403370 Description: iommu: set the default iommu-dma mode as non-strict
Currently, many peripherals are faster than before. For example, the top
speed of the older netcard is 10Gb/s, and now it's more than 25Gb/s. But
when iommu page-table mapping enabled, it's hard to reach the top speed
in strict mode, because of frequently map and unmap operations. In order
to keep abreast of the times, I think it's better to set non-strict as
default.

Below it's our iperf performance data of 25Gb netcard:
strict mode 18-20 Gb/s
non-strict mode 23.5 Gb/s

Team:HISI_SW
Feature or Bugfix:Bugfix

Change-Id: I7695541cd0239b8fdd6945696c09db4e62d1fff8
Signed-off-by: Zhen Lei <thunder.leizhen@huawei.com>
Signed-off-by: x00470154 <x00470154@huawei.com>
Reviewed-on: http://10.90.31.173:8080/5397
Tested-by: public TuringEE <turingee@huawei.com>
Reviewed-by: guohanjun 00470146 <guohanjun@huawei.com>
Reviewed-by: public TuringEE <turingee@huawei.com>