arm_mpam: Summarize feature support during mpam_enable()
To make a decision about whether to expose an mpam class as
a resctrl resource we need to know its overall supported
features and properties.

Once we've probed all the devices, we can walk the tree
and produced overall values. If bitmap properties are mismatched
within a component we cannot support that bitmap.

Signed-off-by: James Morse <james.morse@arm.com>