arm_mpam: resctrl: Pick the caches we will use as resctrl resources
Sytems with MPAM support may have a variety of control types at any
points of their system layout. We can only expose certain types of
control, and only if they exist at  particular locations.

Start with the well-know caches. These have to be depth 2 or 3
and support MPAM's cache portion bitmap controls, with a number
of portions fewer that resctrl's limit.

Hook up resctrl_arch_get_resource(), to do this we need to
generate dummy resources for those that are currently missing.

Signed-off-by: James Morse <james.morse@arm.com>