arm_mpam: resctrl: plumb resctrl configuration through to mpam devices
Because of the smoke and mirrors around the component we picked to
export as resctrl's MBA, the resctrl configuration may be entirely
unlike the mpam configuration that we write into the MSC.

Add a helper to convert from resctrl config to mpam config. This
mpam configuration needs to be applied to a set of devices, not just
any cpu in the resctrl:domain.

Extend the reset logic, to take a specific configuration describing a
partid, and the feature/control to write the mpam config to.

Take all this, an use it to implement resctrl_arch_update_one().

Signed-off-by: James Morse <james.morse@arm.com>