arm_mpam: Pass class and component into mpam_reset_device()
Once an MSC has been reset we should re-apply the resctrl user
configuration. The simplest way to do this is for mpam_reset_device()
to ask for it, as it applies the reset values to the device's
controls.

To do this, mapm_reset_device() needs the component, as that is
where the resctrl configuration lives, and the class, as this has
the control properties that mpam_convert_resctrl_cfg() needs.

Make the mechanical changes to pass these two parameters in.
No change in behaviour.

Signed-off-by: James Morse <james.morse@arm.com>