arm_mpam: resctrl: plumb in cpu hotplug
Currently our mpam cpuhp callbacks take care of the online cpu masks
and resetting/restoring configuration to devices.

Resctrl has a few more requirements, we need to update the domain's
cpu_mask, and later we need somewhere to hook in the resctrl
cpu/domain online/offline calls.

Add the boiler plate to update the resctrl properties of classes
we are exposing as resctrl resources.

Signed-off-by: James Morse <james.morse@arm.com>