arm_mpam: make mpam_component_apply_all()'s machinery use a function pointer
We want to write a monitor configuration to all devices in a class.
The reset logic has the machinery to bounce around the CPUs to get
this done, let's re-use it.

Make mpam_component_apply_all() do its reset work through a function
pointer, so we can supply a different function that updates the monitors.

Signed-off-by: James Morse <james.morse@arm.com>