x86/resctrl: Add arch_has_sparse_bitmaps to explain AMD/Intel CAT difference
Intel expects the cache bitmap provided by user-space to have on a
single span of 1s, whereas AMD can support bitmaps like 0xf00f.

Portable user-sapce has to suck-it-and-see to find out which is
supported. To move resctrl out to /fs/ we need to explain platform
differences like this.

This should be a feature we have to avoid double-negatives when
describing it. Add a resource property arch_has_sparse_bitmaps.
Test this around the 'non-consecutive' test in cbm_validate().

AMD gains the min_cbm_bits test needed for Haswell, but as it
always sets this value to 1, it will never match.

Signed-off-by: James Morse <james.morse@arm.com>