x86/resctrl: Add closid to the staged config
Once we merge the L2/L2CODE/L2DATA resources, we still want to have
two configurations staged for one resource when CDP is enabled.

These two configurations would have different closid as far as the
hardware is concerned.

In preparation, add closid as a staged parameter. In the future this
will be the hardware closid, with the CDP correction already applied
by resctrl. This allows another architecture to work with resctrl,
without having to emulate CDP.

Signed-off-by: James Morse <james.morse@arm.com>