x86/resctrl: Correct the closid when staging configuration changes
Now that apply_config() and update_domains() know the code/data/both value
of what they are writing, and ctrl_val is correctly sized: use the
hardware closid slot, based on the configuration type.

This means we have two 'closids', the single value that maps to an
rdtgroup, and the pair of values that correspond to a configuration
and msr when CDP is use. We call the latter the hw_closid.

Add resctrl_closid_cdp_map() to make one to the other based on
CDP. This means cbm_idx() and its illusionary cache-properties can
go.

Signed-off-by: James Morse <james.morse@arm.com>