x86/resctrl: Walk the resctrl schema list instead of an arch list
Now that resctrl has a list of resources it is using, walk that list
instead of the architectures list. This will let us keep schema
properties with the resource that is using them.

Most users of for_each_alloc_enabled_rdt_resource() are per-schema,
and also want a schema property, like the conf_type. Switch these to
walk the schema list. We only have schema for put alloc_eanbled
resources.

Signed-off-by: James Morse <james.morse@arm.com>