x86/resctrl: Pair up schema when using CDP
When using CDP there are (currently) two resources and therefore
two schema for each piece of hardware. Functions like
rdtgroup_cbm_overlaps() need to take both schema into account when
testing for an overlap.

Add a cdp_peer schema property, and a helper to scan the schema
list to populate it.

This is all temporary, and lets us move the conf_type into
resctrl_schema, (and out of the rdt_hw_resource) before we merge
the configuration arrays and resources. We need two schema as
long as the resource and domains are different.

Signed-off-by: James Morse <james.morse@arm.com>